Changed stars & sizes to badges
Noticed the stars and sizes were outdated. Instead of simply updating, I tought it might be a good idea to have a badge. This way it doesnt lag behind so much. Also moved to the end because I thought the whole table is better readable this way.

Badgen could be replaced with shields.io if need be. I dont know which one is better.